Added features for the second level of homework

+++ backend/ApiService/Source/Application/UseCases/User/Handlers/DeleteUserHandler.cs
@@ -17,16 +17,35 @@ namespace Epam.ItMarathon.ApiService.Application.UseCases.User.Handlers
         public async Task<Result<RoomAggregate, ValidationResult>> Handle(DeleteUserRequest request,
             CancellationToken cancellationToken)
         {
-            // 1. Get room by user code
+            // 1. If UserCode not found, return BadRequestError
+            if (string.IsNullOrEmpty(request.UserCode))
+            {
+                return Result.Failure<RoomAggregate, ValidationResult>(new BadRequestError([
+                    new ValidationFailure("userCode", "userCode is required.")
+                ]));
+            }
+
+
+            // 3. Get room by user code
             var roomResult = await roomRepository.GetByUserCodeAsync(request.UserCode, cancellationToken);
 
-            // 2. If room not found, return NotFoundError
+
+            // 4. If room not found, return NotFoundError
             if (roomResult.IsFailure)
             {
                 return roomResult;
             }
 
-            // 3. If room found, delete user by id from room
+            // 2. If userCode and UserID is the same user, return BadRequestError
+            var user = roomResult.Value;
+            if (user.Id == request.UserId.Value)
+            {
+                return Result.Failure<RoomAggregate, ValidationResult>(new BadRequestError([
+                    new ValidationFailure(string.Empty, "User cannot delete himself.")
+                ]));
+            }
+
+            // 5. If room found, delete user by id from room
             var room = roomResult.Value;
             var deleteResult = room.DeleteUser(request.UserId);
             if (deleteResult.IsFailure)
@@ -34,7 +53,7 @@ namespace Epam.ItMarathon.ApiService.Application.UseCases.User.Handlers
                 return deleteResult;
             }
 
-            // 4. Update room in repository
+            // 6. Update room in repository
             var updateResult = await roomRepository.UpdateAsync(room, cancellationToken);
             if (updateResult.IsFailure)
             {
@@ -43,7 +62,7 @@ namespace Epam.ItMarathon.ApiService.Application.UseCases.User.Handlers
                 ]));
             }
 
-            // 5. Get updated room
+            // 7. Get updated room
             var updatedRoomResult = await roomRepository.GetByUserCodeAsync(request.UserCode, cancellationToken);
 
             return updatedRoomResult;
